Add atomic implementation
- move memtier structs into memkind_memtier.c
- detect possible atomic operations
- provide increment/decrement atomic operations
