+++ src/memkind_memtier.c
@@ -6,7 +6,54 @@
 #include <memkind/internal/memkind_arena.h>
 #include <memkind/internal/memkind_log.h>
 
+#include "config.h"
 #include <assert.h>
+
+#ifdef HAVE_STDATOMIC_H
+#include <stdatomic.h>
+#define MEMKIND_ATOMIC _Atomic
+#else
+#define MEMKIND_ATOMIC
+#endif
+
+#if defined(MEMKIND_ATOMIC_C11_SUPPORT)
+#define memkind_atomic_increment(counter, val)                                 \
+    atomic_fetch_add_explicit(&counter, val, memory_order_relaxed)
+#define memkind_atomic_decrement(counter, val)                                 \
+    atomic_fetch_sub_explicit(&counter, val, memory_order_relaxed)
+#elif defined(MEMKIND_ATOMIC_BUILTINS_SUPPORT)
+#define memkind_atomic_increment(counter, val)                                 \
+    __atomic_add_fetch(&counter, val, __ATOMIC_RELAXED)
+#define memkind_atomic_decrement(counter, val)                                 \
+    __atomic_sub_fetch(&counter, val, __ATOMIC_RELAXED)
+#elif defined(MEMKIND_ATOMIC_SYNC_SUPPORT)
+#define memkind_atomic_increment(counter, val)                                 \
+    __sync_add_and_fetch(&counter, val)
+#define memkind_atomic_decrement(counter, val)                                 \
+    __sync_sub_and_fetch(&counter, val)
+#else
+#error "Missing atomic implementation."
+#endif
+
+struct memtier_tier {
+    memkind_t kind;                   // Memory kind
+    MEMKIND_ATOMIC size_t alloc_size; // Allocated size
+};
+
+struct memtier_tier_cfg {
+    struct memtier_tier *tier; // Memory tier
+    unsigned tier_ratio;       // Memory tier ratio
+};
+
+struct memtier_builder {
+    unsigned size;                // Number of memory tiers
+    unsigned policy;              // Tiering policy
+    struct memtier_tier_cfg *cfg; // Memory Tier configuration
+};
+
+struct memtier_kind {
+    struct memtier_builder *builder; // Tiering kind configuration
+};
 
 // Provide translation from memkind_t to memtier_t
 // memkind_t partition id -> memtier tier
@@ -175,7 +222,7 @@ MEMKIND_EXPORT void *memtier_kind_malloc(struct memtier_kind *kind, size_t size)
 MEMKIND_EXPORT void *memtier_tier_malloc(struct memtier_tier *tier, size_t size)
 {
     void *ptr = memkind_malloc(tier->kind, size);
-    tier->alloc_size += jemk_malloc_usable_size(ptr);
+    memkind_atomic_increment(tier->alloc_size, jemk_malloc_usable_size(ptr));
     return ptr;
 }
 
@@ -189,7 +236,7 @@ MEMKIND_EXPORT void *memtier_tier_calloc(struct memtier_tier *tier, size_t num,
                                          size_t size)
 {
     void *ptr = memkind_calloc(tier->kind, num, size);
-    tier->alloc_size += jemk_malloc_usable_size(ptr);
+    memkind_atomic_increment(tier->alloc_size, jemk_malloc_usable_size(ptr));
     return ptr;
 }
 
@@ -210,17 +257,21 @@ MEMKIND_EXPORT void *memtier_tier_realloc(struct memtier_tier *tier, void *ptr,
                                           size_t size)
 {
     if (size == 0 && ptr != NULL) {
-        tier->alloc_size -= jemk_malloc_usable_size(ptr);
+        memkind_atomic_decrement(tier->alloc_size,
+                                 jemk_malloc_usable_size(ptr));
         memkind_free(tier->kind, ptr);
         return NULL;
     } else if (ptr == NULL) {
         void *n_ptr = memkind_malloc(tier->kind, size);
-        tier->alloc_size += jemk_malloc_usable_size(n_ptr);
+        memkind_atomic_increment(tier->alloc_size,
+                                 jemk_malloc_usable_size(n_ptr));
         return n_ptr;
     } else {
-        tier->alloc_size -= jemk_malloc_usable_size(ptr);
+        memkind_atomic_decrement(tier->alloc_size,
+                                 jemk_malloc_usable_size(ptr));
         void *n_ptr = memkind_realloc(tier->kind, ptr, size);
-        tier->alloc_size += jemk_malloc_usable_size(n_ptr);
+        memkind_atomic_increment(tier->alloc_size,
+                                 jemk_malloc_usable_size(n_ptr));
         return n_ptr;
     }
 }
@@ -237,7 +288,8 @@ MEMKIND_EXPORT int memtier_tier_posix_memalign(struct memtier_tier *tier,
                                                size_t size)
 {
     int res = memkind_posix_memalign(tier->kind, memptr, alignment, size);
-    tier->alloc_size += jemk_malloc_usable_size(*memptr);
+    memkind_atomic_increment(tier->alloc_size,
+                             jemk_malloc_usable_size(*memptr));
     return res;
 }
 
@@ -252,7 +304,7 @@ MEMKIND_EXPORT void memtier_free(void *ptr)
     if (!kind)
         return;
     struct memtier_tier *tier = memtier_registry_g.kind_map[kind->partition];
-    tier->alloc_size -= jemk_malloc_usable_size(ptr);
+    memkind_atomic_decrement(tier->alloc_size, jemk_malloc_usable_size(ptr));
     memkind_free(kind, ptr);
 }
